pyproject.toml
[project]
name = "ml-agent-assistant"
version = "0.1.0"
description = "Multi-Agent Data Science & ML Assistant"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "langsmith>=0.1.0",
    "e2b-code-interpreter>=1.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "streamlit>=1.30.0",
    "docker>=7.0.0",
    "pydantic>=2.0.0",
]
